% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{DownloadURLList}
\alias{DownloadURLList}
\title{Download Files from URL List}
\usage{
DownloadURLList(path)
}
\arguments{
\item{path}{Path to the file containing URLs}
}
\value{
A string indicating the result of the download process
}
\description{
Reads a file containing a list of URLs (one per line) and downloads the files.
}
\details{
This function downloads files from URLs listed in a file.
}
\examples{
\dontrun{
DownloadURLList("/path/to/url_list.txt")
}
}
