% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{DownloadZoteroPDFs}
\alias{DownloadZoteroPDFs}
\title{Download PDFs from Zotero}
\usage{
DownloadZoteroPDFs(username, api_key, collection_name, parent_dir)
}
\arguments{
\item{username}{Your Zotero username/user ID}

\item{api_key}{Your Zotero API key}

\item{collection_name}{The name of the Zotero collection}

\item{parent_dir}{Directory where PDFs will be saved}
}
\value{
A string indicating the result of the download process
}
\description{
Downloads PDF documents from a Zotero collection using the Zotero API.
}
\details{
This function downloads PDFs from a Zotero collection to a specified directory.
}
\examples{
\dontrun{
DownloadZoteroPDFs("user123", "apikey456", "My Collection", "/path/to/pdfs")
}
}
